refactor: move AzureIdentityLogForwarderHostedService into Logging/
Code review flagged this as belonging with the other logging-infrastructure
classes (FileLoggerProvider, NamedPipeLoggerProvider, etc.) in
Application.Common/Logging/, per this repo's own file-placement rule —
Extensions/ is for DI extension methods, not the services they register.
