+++ src/Content/Application/Application.Common/Extensions/AzureIdentityDiagnosticsExtensions.cs
@@ -1,3 +1,4 @@
+using Application.Common.Logging;
 using Microsoft.Extensions.Azure;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
